Gui changes:  - Added kinetic scrolling to switch on the fly  - Base font size can be changed in the GUI
Font size changes are not perfect yet for oculars and satellites,
and the GUI panels also depend on GUI basefont size.
This should be separated.

+++ src/core/modules/SolarSystem.hpp
@@ -422,7 +422,7 @@ public slots:
 
 	//! Set planet names font size.
 	//! @return font size
-	void setFontSize(float newFontSize);
+	void setFontSize(int newFontSize);
 
 	//! Set the color used to draw planet labels.
 	//! @param c The color of the planet labels (R,G,B)
